investigation and new structure discovery
